Software/Subscraper/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
chromedriver-autoinstaller==0.6.2
pyvirtualdisplay==3.0